Divided the DocumentationGen into DocumentationGenHtml & DocumentationMarkup
